Fixed typos.

+++ squawkbox/modules/positional_embedding.py
@@ -12,7 +12,7 @@ class PositionalEmbedding(torch.nn.Module):
         self.freq = 1 / 10000 ** ( torch.arange(0, half_dim, dtype=torch.float32).view(1, 1, half_dim) / half_dim)
 
     def forward(self, timestamp):
-        timestamp = timestamp.unsqueeze(-1)
+        timestamp = timestamp.type(torch.float32).unsqueeze(-1)
         x = torch.matmul(timestamp, self.freq)  # (batch_size, seq_len, dim / 2)
         sin = torch.sin(x)
         cos = torch.cos(x)
